File Organizing, Repository Updates, and Prep Work
Rearranged some files and updated our formatting system. Files will be kebab-case, whereas folders will be snake_case.

Updated the name of the repository to be FoodWorld rather than 2024Game.

Created a new Godot project that I'll be working out of for the duration of this game and set up some of the files for it.