Player Node and Script in Godot
I began working on the Player Node and script in Godot.

For the Node I've added a Camera2D, AnimatedSprite2D, CollisionShape2D, and two timers for dodging. I also added a TileLayer into the Game Scene using Ryan's grass tile.

In the script, I added support for movement- including sprinting and dodging. I also have all of the variables prepared for HP, XP, and stamina. I was able to implement the different fight styles when you press 1, 2, or 3 on the keyboard, and it works great! Once we have more assets I'll add animations to go with the keypresses. It's good to have this system set up as it will be critical to the game because you will often need to swap out Food buddies.

Before our session on Friday 9/18 I'd like to implement stamina that depletes and  regens over time, finish up sprinting and dodging functionality by adding camera zoom effects when those actions occur, allow dodges to occur while standing still, functionality for jumping, and the backend for inventory (not the visual graphic for it).

After that, I'd like to work on player attacking logic and player taking damage logic (no animations for this yet, just get it so that inbound and outbound damage with respect to the player properly registers). Once that's complete, the vast majority of the Player's functionality will be implemented and I will probably add a few things to it here and there. I can probably move onto another script or Node after the attack logic is complete.