Player: Jump Functionality, Inventory Variables, Comments
I added two variables, one for the inventory array and one that defines a max size for the inventory. Nothing else being done with these yet, and we might not end up doing much at all with it. We need to discuss what we want for items in the game and see if there are ways we can integrate them in well.

I tried working on jump functionality, and I've got some of it working, but it is finnicky because of the top-down perspective. I tried making the jump bring you back to the initial point you jumped from after gravity is applied, so that you don't fall down and off the map, but it's not working 100% and won't be a permanent solution anyways. We might not move forward with putting jumping into the game, but that depends on if we think of ways that jumping would be useful/necessary in-game. Since we'll be relying on Food Buddies and fusions primarily for world traversal and exploring, we might not need the Player to have a jump ability.

Also went through and commented things so that they're easier to understand when I come back.