Player: Jump Functionality Improvement
I made some adjustments to improve the jump functionality a bit more. It is MUCH more stable than in the previous commit for jump functionality, because now the player will always land at the same y-coordinate that they initiated their jump from and they must press the jump key manually each time rather than holding down on it.

It is not 100% functional yet, because this only allows the player to jump up and down on level ground. They wouldn't be able to jump on top of things or jump and fall off of ledges because the jump ends once the player falls back to the y-coordinate they initiated the jump from.