Enemy & Food Buddy Nodes, New Signals, Ability Logic, RigidBodies, and more
I created scenes for an Enemy and a Food Buddy. I have attached scripts to both and have begun working on them. I connected signals from the Player script to the Enemy script that will trigger whenever the Player attacks solo (left-click or right-click), and the Enemy has a callback function to process the Player's attack. This is functional and the enemy takes damage when the Player punches or kicks within the enemy hitbox.

I also have signals connected from the Player script to the Food Buddy script that will trigger whenever the Player attacks using one of the two Food Buddies. I added variables in the Player script to keep track of the two active Food Buddies and a whole list of Food Buddies that the Player has acquired so far. The Food Buddy script has a callback function to process this signal and use their ability.

I made some changes to the stamina code. I added two functions: one that is good for depleting stamina once, and another that is good for depleting stamina over time with delta. I also made it so that no checks (like is_sprinting or is_dodging) are being made in the update_stamina function.

I experimented a bit with RigidBody2D, which is what Enemy and Food Buddy have set as their root node. I created a function that will push a RigidBody2D when the Player collides with them, but I currently have it disabled because I need to do further experimenting and tweaking with it to understand how to implement physics the way I'd like to.

Went through and organized my code as best I could.