Player, Food Buddy, and Enemy Ability Signal Work & Updated Tilemap
I worked on the signals that will help the Player, Food Buddy, and Enemy communicate with one another. I condensed the number of signals being used by the Player and modified some code so that it was more dynamic and multi-purpose. I also needed to update the enemy's callback functions that were processing ability use and attacks.

I added new signals and variables to the Food Buddy and began working on the Food Buddy's solo attack behavior. It's actually coming along pretty nicely, but the hard part is going to be making it automatically move towards an enemy. I'm trying to design this in a way that each Food Buddy will be able to process ability uses the same, despite having different abilities. As long as I can connect the Food Buddy to the Player and Enemy nodes okay, I can worry about individualistic ability behaviors in callback functions later on.

I also revamped the tilemap so that some of Ryan's new tiles were able to be used, but it seems like there's an issue with it. I looked further into it and realized that the opacity on some of the edges of the tiles was reduced, and I think that was from copying and pasting them from their individual files into the mega spritesheet. I'll have Ryan fix that tomorrow.