Food Buddy, Enemy, and Game Work (Node Changes, Attacking Logic, Loading Scenes, Signal Connections, and more)
Swapped Food Buddy and Enemy from RigidBody2Ds into CharacterBody2Ds because it seems like that will be the more appropriate Node for moving them. For some reason I was assuming that the CharacterBody2D was specifically for user-input, but that isn't the case.

I revamped some of the Node variables in the food_buddy.gd file so that they are able to be easily updated and reassigned when a subclass is made.

Discovered a 1-liner for loading in assets such that their logic and their Nodes are all loaded into the game. You can see it in the game.gd file, line 7. Then I have code in the process function starting at line 34 that shows how to load duplicates of the instantiated sausage link scene. You can also load in the regular one, but my plan for this is to only need to load in the resource in code once, then I can make duplicates of that resource that will be added/removed as I please from the main scene tree with the game as the root node.

Got the Food Buddy solo attack working. Enemies die when Food Buddies deal enough damage, then the Food Buddy will choose the next closest on-screen target.

I had to manually connect signals through code as the sausage link scene was loaded in because it doesn't seem like I can connect to a signal through inheritance of Food Buddy or through the Godot editor unless I place sausage link in the game scenetree from the start.

I organized all of the code so that it is well-commented and clear to work with.