Enemy Moving & Attacking Functionality + More
Food Buddy Script Updates:
Added the "Forage" FieldState and an inventory into the FoodBuddy script. This state will direct the FoodBuddy towards any berries to collect and store in their inventory. Later on I can add functionality so that when the Food Buddy's inventory is filled or when there are no more berries on the screen, it can start following the Player again. Once it is in-range of the Player, the berries it collected can jump out of its inventory and enter the portable sauna.
I revamped some of the signals and variables in the Food Buddy script so that they were more multi-purpose. Instead of move_towards_player and move_towards_enemy, I now only have move_towards_target. I added new signals for target_player and target_closest_enemy which when emitted will have the Game set the desired target for the Food Buddy, and then when move_towards_target is emitted the Game script moves them towards the currently selected target.

Enemy Script Updates:
I created Enemy signals, variables, and functions. I got the Enemy to a similar state that the Food Buddy is in by revamping the "use_solo_attack" function from the FoodBuddy script in the Enemy script. After connecting Enemy signals to the game and reworking that function, the Enemy can now automatically move towards either the Player or the Food Buddy and is able to deal damage to and kill both of them.

Player Script Updates:
Added "die" and "killed_target" signals. I also discovered some stuff about physics collisions with CharacterBody2D but didn't have time to look into that yet, so I saved a link above the check_collide_and_push function.

Game Script Updates:
Capitalized the variable names for the references I have to the Player, Food Buddy, and Enemy nodes because they are constant. GDScript doesn't let me label them as const rather than var because I have the @onready annotation in front. If I wanted to I could work around this.
Renamed lots of variables in functions from "enemy" to "target" so that the functions can be more multipurpose.
Added a callback function for the Player's "die" signal and currently have it remove the Player from the SceneTree. Eventually I'll change this so that maybe the Food Buddy can go revive the Player when they're dead.
Added all of the Enemy callback functions and connected all of the signals.

Everything is commented and documented as it should be.