Quick Bug Fixes: Enemy Attack Persisting & Food Buddy Not Following Enemy
The Enemy was still attacking even after the Food Buddy was removed as a child, and I realized that was because although it had been removed from the Game's SceneTree, the Player still held a reference to it and that was being detected through the Enemy code. Once the Food Buddy dies now, a check is made to see if that Food Buddy was one of the Player's active ones, and then it is set to null in the Player's inventory.

The Food Buddy started following the Player even while in the FIGHT field state. This only occurred after I toggled it's field state from FIGHT, to FOLLOW, then back to FIGHT. This was because when the FOLLOW state occurs, the Food Buddy begins to target the Player, and when it swapped back to FIGHT it wasn't changing it's target back to an enemy. Now I have it working as intended.