Player Food Buddy and Fusion Equip/Unequip Logic
I added a second Food Buddy Node into the Game Scene in Godot as I'm now trying to test switching FieldStates for the Player and the Food Buddies with both Food Buddies active at once. It is getting there, but not quite complete yet. I need to add logic to switch Food Buddy 1 back to its previous state if it  is in the PLAYER state, but then the Player equips Food Buddy 2.

I changed the animation names in the Player Node in Godot to no longer include "fighting". I want FieldState to be a universal descriptor, and having FightStates was unnecessary when I can just store all of the different state types in the FieldState enum.

I disabled the Enemy's targeting in enemy.gd so it currently isn't attacking the Player or any Food Buddy.

I removed all FightStyle variables and related code for Food Buddies in food_buddy.gd. I expanded on the FieldState variables in the place of FightStyle ones. I also switched AttackDamage from being an enum to being a dictionary (like I had done for player.gd a few commits back), because enums are constant but attack damage for each Food Buddy won't be. I also had to increase the target_distance range while the Food Buddy is trying to attack solo, because I discovered that the registration point of Nodes is on the bottom left of the Node. When the Food Buddy was attacking from the right side while the range of execution for the solo ability was 10 px, the Food Buddy wasn't emitting the signal because even while touching the Enemy, the Food Buddy was still technically more than 10 px away from the bottom left registration point. I also added some debugging print statements to display when the Food Buddy's abilities are being triggered by the Player.

I added the 'killed_target' and 'die' signals to the food_buddy_fusion.gd file, but didn't do anything with them yet.

I updated the Node variables in game.gd to include two Food Buddies.
I decided to keep track of active Food Buddies in the game.gd file rather than in the player.gd file so that more of the logic is concentrated into this one file. I added an array to store which two Food Buddies are active, so I was able to remove the 'get_active_food_buddies' function because the Game will always have access to them now.
I also added an array to store which unlocked Food Buddies aren't currently active, and another array to store which Food Buddies the Player hasn't unlocked yet. I made these variables with a Food Buddy selection interface in mind (which I think should be a quick easy to interact with menu that is triggered when the 'R' key is pressed). I also made similar variables for Food Buddy Fusions to be tracked.
I added the two Food Buddy Nodes into the active Food Buddies list in the game.gd '_ready' function to begin testing on them.
I removed the required array type of Node2D for the 'targets' parameter in the 'select_closest_target' function because for some reason it wouldn't classify FoodBuddy as a Node2D through inheritance/polymorphism despite FoodBuddy extending CharacterBody2D, which extends from Node2D. Now it just requires any Array.
I added an "equip_food_buddy_fusion" function which will iterate over a list of unlocked Food Buddy Fusions and determine if your two Food Buddies would comprise one of those fusions, then it sets that fusion as the currently active one and returns it.
I added logic to Player callbacks for equipping buddies or buddy fusions.

I modified the names of the "equip_buddy" and "equip_buddy_fusion" signals so that they'd say "toggle" rather than straight up "equip", because those signals are multipurpose and can equip or unequip.
I removed references to Food Buddies from the Player, so when they try to equip one they just emit the Food Buddy's number that they're trying to equip and then the game processes that number to determine which Food Buddy the Player wants.
I also updated the FieldStates for the Player similarly to the Food Buddy script. No more FightStyle.