Player Equip/Unequip Logic & Added Real Food Buddies
I worked on making sure the equipping/unequipping logic for Food Buddies works properly. Previously it was allowing for two Food Buddies to remain in the Player state at once, which shouldn't occur. Now it only allows one.

Implemented new nodes for Malick and Sally to visualize scale and test, and they look good proportionally in the game. There seems to be an unexpected error with signals though, and I'll have to find a clever way to get it to work the way I need it to. I might need to define a parent node that the Food Buddy Scene's signals connect to, then when I bring it into the Game Scene it will be connected, so I'll just have to connect the Food Buddy parent node's signals to the Game so that they can process properly.