Malick and Sally & FieldState Interface Testing
Added Malick and Sally into the Game SceneTree and removed the blueprint Food Buddies. I figured out how to connect the signals to the game so they are functioning now. I just had to specify that they should connect to the "_on_food_buddy" callbacks, not "_on_malick" or "_on_sally", because that would create new callbacks entirely.

I implemented code for the Food Buddy FieldState Interface functionality through a callback in game.gd, which responds to the trigger signal is coded player.gd. It's actually
working really well. You can use 'Q' to open the menu. It is currently displaying through the console with print statements. When the menu is active, all Food Buddies and enemies stop processing entirely, and the Player's 'pause' variable is set to true.
You can use 'A' and 'D' to switch between the selected Food Buddy, and then you can use 'W' and 'S' to switch between the selected FieldState for the currently selected Food Buddy. The selection options include all unlocked FieldStates except for Fusion (sort of hardcoded to prevent Fusion from being selected because I set Fusion as the final FieldState and avoid it ever being selected like that in game.gd _process function). Fusion should only be triggered by pressing '3' (if we allow it to be selected any other way, there could be a situation where one Food Buddy is in FUSION mode but the other isn't, which shouldn't be possible). Additionally, pressing '1', '2', and '3', also work in this menu as the typical shorthands for Player-Buddy1, Player-Buddy2, and Player-Fusion respectively.
Pressing 'Q' again will close the interface and resume all of the processing for the characters that were paused.

I made sure that the interactions with platform layers for any CharacterBody2D nodes that I have are disabled, since those are designed for sidescroller games and don't work well for top-down games. It was causing an issue where if the bottom of a character touched another character, it'd get stuck because it'd treat the other character like a platform.

Added "alive" variables to each CharacterBody2D (Enemy, Player, Food Buddy, and Food Buddy Fusion). There was an issue in situations where an Enemy being targeted by two Food Buddies would still remain the target of whichever Food Buddy didn't kill it, and that was because the process_attack( ) function in game.gd only sets the target to null for whoever killed it, not anyone who is targeting it. The alive variable was an easy fix for this, because then the Food Buddy can switch their target when their current target.alive == false.