Start of Dialogue Interface
I am carefully thinking out a good design for the Dialogue system. I'm trying to design a system that will work with the dialogue in "conversations", so certain scenes of dialogue are saved in separate files or objects.

With the "conversations" idea, I can find a way to simply pass the game a "conversation" file/object and then I can base the interface design and controls on how conversations are to be processed.

I need to consider the different kinds of conversations and dialogue that will occur. Sometimes the Player may be prompted to select a dialogue option. Sometimes the dialogue that occurs will be based on the current events of the game or the Player's progress in the game. I need a system that can handle any and all kinds of dialogue.

I currently imagine the "conversation" object/file would be processed line by line by the game. Each line would store the actual text that should be displayed, the person/food buddy/enemy/citizen that is speaking, and access to their animated sprite.