Dialogue System Work: Foundation Secured
I did a lot of thinking for the structure of the Dialogue system, and I think I've come up with a good way to build it. I typed out a description of it below and it is saved in the Game Dev Google Doc. I began implementing part of the Dialogue Resource today after thinking about the structure for about 2 hours. I figured out how I'd be storing names of Dialogue participants and their lines. I have the names of the Dialogue participants as the key to a dictionary, and the value for each of those character names is another dictionary. That dictionary contains a number as its key, which represents what the number of the line is and the order it should be played in the Dialogue. The value for that key is the actual line, the String. I have added code for adjusting the line of dialogue currently being displayed in a simple function that allows the Player to move forwards and backwards through the dialogue without any errors.
A system where each possible conversation between a specific group of characters is stored in one specified file for that unique group of characters. That means if there were a total of 57 conversations between the Player and Sausage Link, they'd all be stored in a single resource specific to those two. If there were 84 conversations between the Player, Donut, Sausage Link, and Squid, they'd all be stored in a single resource specific to those 4.

When the Player presses 'E' to interact and engage in a conversation with an NPC, the Dialogue Interface should be triggered. When the signal to activate the interface is emitted from the Player, the Game should have a callback function to process it. The Game will have logic to determine which dialogue resource should be loaded in by considering who is involved in the dialogue. When the Player triggers this dialogue with an NPC, the Game should consider the currently active Food Buddies, the NPC that the Player has engaged an interaction with, and whether or not the interaction is going to progress the story of the game.
After processing all of that information (and likely much more information that I think of later), the Game will generate a String that contains the name of each participant in the dialogue. For example, if the Player has interacted with Link, there should be code that takes the names of each conversation participant and loads in the resource with that name, like: "load('Player-Link')". The Game can also have logic to know what order to place each name in, for example, like knowing that the Player is always first, then Food Buddies (organized alphabetically), then enemies (alphabetically), then NPCs (alphabetically), etc. so that there are no problems with loading the right resource. Errors can be easily caught because if I try to load a resource that doesn't exist, it will just set the variable I loaded the resource into to 'null', so I can easily check to make sure it isn't null before putting that dialogue into the interface.

If the Player was speaking 1-on-1 to an NPC named Gary and the Food Buddies won't speak, the file name should be "Player-Gary"
If the Player was speaking to Gary again but this time the Food Buddies would be involved, the file name will be different. If specific Food Buddies would have a specific effect on the outcome of the dialogue, the file should be named like "Player-Link-Dan-Gary".
If only Link affects that dialogue and the other Food Buddy doesn't, the file should be named "Player-Link-FoodBuddy-Gary".
I can make it so that each Food Buddy has their own way of saying what would really essentially be the same thing. For example, if they were excited Link might say "Hell yeah" but Dan might say "BOOYAH!!", both conveying their excitement in their own ways based on their own personalities.
With the idea to add this 'FoodBuddy' filler tag into the file-naming scheme, I'd need a way to maintain those personalities even though I might not specifically know which Food Buddy is in the conversation.
In the instance that one of the files has the 'FoodBuddy' tag in it, I will add dialogue lines for every possible Food Buddy into the dialogue resource that each essentially say the same thing and steer the dialogue the same way, but I'll make sure they reflect the personality of whichever FoodBuddy is saying it.

Once the resource is loaded as the "current dialogue" for the Dialogue Interface, the interface should be able to call functions in response to user input on the "current dialogue" resource that progress the conversation forward.

There could also be other more complex interfaces created for dialogue that will occur mid-battle, or dialogue that will occur in an uninteractable cutscene format, or dialogue that will occur in an interactable cutscene format! Each of these could use a different. The default dialogue resource works off of the "dialogue.gd" script, but there could also be extensions off of that foundation such as a "dialogue_cutscene_uninteractable.gd" script. This would be a subclass of the "dialogue.gd" script, and with that new child script a different kind of dialogue resource can be created that will handle an automatic progression of the dialogue since it is a cutscene that won't take Player interaction.