More Dialogue Thinking & Structuring
I put more time and thought into the structure of the Dialogue system. My goal is to make it be as dynamic and flexible as possible so that it will be very easy and simple to use. I decided it'd make most sense for Ryan and I to type out the dialogue for each character in a format that is similar to how scripts are written for plays. I spent time thinking of ways I could structure these "scripts" so that they could contain not just dialogue, but also indicate what animation should play as a character says a line, indicate how the text should be displaying (big and shaky for anger and aggression, small and quivering for scared and shy, etc.). I think we can find a consistent format to type the scripts in that will feel natural and won't slow us down as we're coming up with dialogue while also allowing me to pass that script into my code and have it process the script into a dialogue resource, which will then be passed in to the interface.

I modified the dialogue.gd file so that it has an array called 'conversations' to store every conversation that exists between the characters of the file. For example, if the Player and Sausage Link had 50 conversations throughout the whole game, I would type out each of those conversations in one .txt file (adhering to the format that the code recognizes and can parse), then I'd pass the file into the algorithm to be parsed. The algorithm will know when a conversation starts and ends, so it will be able to create a 'conversation', store it in the 'conversations' array, and then it will begin creating the next one. The end result would be 50 distinct conversations stored in the 'conversations' array.
I also made some revisions to the adjust_current_line function so that it is easier to read and understand.
I started making the 'create_new_dialogue_resource' function. This is going to a function that I use to generate each dialogue resource file. Once I have all of the dialogue resource files created, this function won't be used anywhere throughout the game- it is solely to get my .txt script for the dialogue translated into code that can be executed and processed by the dialogue.gd file as well as the Dialogue Interface.

The game.gd file currently has experimental code with the 'testtext.txt' file. I wanted to see how the FileAccess system in Godot works, and fortunately its very similar to Python (if not better!). I plan to structure the .txt script files in a manner that they can easily be parsed with the 'get_line( )', 'rstrip( )', and potentially the 'rstrip( )' functions. I'll use delimiters as I type the scripts so the algorithm will have a good reference on where to split the code.