Dialogue Interface Comments
game.gd:
Added comments to the callback function for triggering the Dialogue Interface.
Refactored interface variable names so that they all start with the type of interface. Even though this makes the names much longer, I think it is best for clarity.
Added variables for the Dialogue Interface and began working on the processing code.