Dialogue Interface Processing
I worked on processing a Dialogue Resource through the Dialogue Interface. I haven't completed the code for parsing a .txt file and creating a Dialogue Resource, so I am working with one that I built up manually through Godot. I also didn't initiate the interface through the 'toggle_dialogue_interface' callback function because it isn't complete yet, so I'm manually assigning all of the values and triggering the interface through the game's '_ready' function.

dialogue.tres and dialogue.gd:
I added in some test dialogue and worked with the data structure of how to store conversations. All conversations are stored in a hierarchy of dictionaries.

game.gd:
I added lots of code that is pretty messy right now because I'm manually triggering the Dialogue Interface, but when this is ready to be triggered by Player input, all of that code in the '_ready' function for the Dialogue Interface will live in the 'toggle_dialogue_interface' callback function.

malick.gd and player.gd
Added statements in both '_ready' functions to set their names. These names are read in through the Dialogue Interface and end up being important for matching data from the Resource to the Game.