Dialogue Parser & Resource Creator Complete
dialogue.gd
I created the "create_and_save_resource( )" function to handle the parsing of the Dialogue scripts I will be typing. It took some time, but it is working smoothly and (surprisingly) exactly as I planned! I added detailed comments, probably more than necessary, but I figured this was a very specific extraction of data so it deserved a very specific set of comments to help me remember why things are being done the way they are.
The function works. It parses the .txt file and creates conversations out of the Dialogue and stores them in the Resource, then it saves into a brand new separate Resource.

Next I just need to test that it'll play nice in the interface, which might not take long, but I'm pretty wiped right now and it's 2 AM so I'm tapping out here. Supposed to have game dev with Ryan tomorrow so I'll tackle it then.