Dialogue Parsed Resource Works In Interface
game.gd
I used a Dialogue Resource created from a parsed .txt file in the Dialogue Interface. It worked and processed okay, but I had to add some foundational support for the GAME and DIALOGUE directions that come from the .txt file. These will need to be further expanded on to allow dynamic support for necessary Game and Dialogue updates, but are in a pretty good place right now.