Trigger Dialogue "Press 'E' To Interact" Functionality Complete
food_citizen.gd:
I built up the foundation for the Food Citizen class using the Enemy class as a reference. The Food Citizen can target the Player or Food Buddies and holds a Label node. The Label has text that says 'Press 'E' To Interact'

dialogue.gd:
I added a 'prepare_dialogue' function which sets up a conversation to be played in the Dialogue Interface. It sets the current conversation based on a given conversation name as well as the first speaker, the line, and the line number the dialogue should start on.

player.gd:
Added signals for 'interact' and 'escape_menu'
Added emit statements for the two signals in the _process function. The 'interact' signal can't be emitted while the Player is paused, but the 'escape_menu'  signal can be.

game.gd:
I condensed similar callback functions for the Food Buddy and Enemy into "_on_character..." callbacks. I reconnected all of the signals through the Godot Editor.
Added a function that will make the Food Citizen's label appear/disappear dependent on if the Player is in range of them to start a conversation. This function is very geared towards this one specific test Food Citizen, but I plan to make it much more dynamic and applicable to various other NPCs.
Added a callback function to process a Player's pressing of the 'E' interact key. This function will trigger a conversation in the Dialogue Interface if the test Food Citizen's Label is visible (player is in range). I plan to make this function capable of checking various NPCs rather than just one, and I also plan for it to be able to determine the specific conversation to play.
Added a callback function to process the Player's pressing of the 'ESC' escape menu key. This callback simply disables the Dialogue Interface if the Player is actively in it when ESC is pressed.
I split up the "toggle_dialogue_interface" function into two enable/disable functions.