Code & File Restructuring, Full System Testing
I rearranged a ton of code, particularly the two interfaces for Dialogue and Food Buddy FieldState have been moved out of the game.gd and into their own scripts and Scenes. This structuring will be easier for when I begin to add UI elements that will likely need to be stored in Nodes within Scenes.

I also made a new Character script and Scene for Enemy, Food Buddy, Food Citizen, and Player to inherit from. This seems to be working well, and it will make code updates a lot easier to manage since it's all majorly isolated into this Character now.

I switched Characters to use center points as their registration point and that was a bit of a struggle at first but seems to be working well now.

I probably did a bunch more, but I can't remember. Planning to flesh out the 'Press 'E' To Interact' logic and make it more dynamic rather than just hard-coded for the one Food Citizen as it is currently. After doing that, I'll work on the logic for determining who is involved in a conversation and I'll handle the automatic conversation selection logic for Dialogue.