Interaction System: Basic Foundation Complete
I have added support for the Player to be able to interact with multiple entities, not just the single Food Citizen I had initially tested with. The Player can now interact with anything that is classified as an "Interactable", and that classification was made when I set certain Nodes within the Godot Editor to be a part of an 'interactables' group. Food Buddies and Food Buddies are currently the only two entities I have for testing as Interactables, but various different Assets can be Interactables too (items, doors, map assets, etc.)

The next step is for each Interactable to have a function that will define the logic that the Interactable should follow when the Player interacts with them. I think I'm going to make an 'interactable.gd' Node and script that will extend from the Character script.