Dialogue Basic Conversation Selecting Logic
I made different scripts for when Player and Sally 1-on-1 talks and then Player and Malick 1-on-1 talks. I wrote the code in a way that consider which Food Buddies are in-range of the Player. If both, it includes both in the conversation. If only 1 is in range, it only includes the one that you're talking to because the other one isn't close enough to be part of the conversation.

This needs a bit of work, and the code for selecting conversations needs to be expanded upon. I want to wait to do this until we've written more Dialogue and have a better idea of how we'll be organizing it, then I'll write more advanced code for selecting conversations.