Bug Fixes and Code Comments
I fixed a bug with the parser algorithm in dialogue_resource.gd. The algorithm wasn't parsing the final line of dialogue in a conversation, so I simply added one more line to the end of each Google Doc .txt file that says "END", and that fixed it.

There was also another bug where if a DIALOGUE or GAME instruction is the last line of the file, then it won't process the conversation properly and continually reprint the instruction. The only reason this bug presented itself was because a DIALOGUE instruction was the second to last line of the file but the line after it wasn't parsed and saved by the algorithm. I don't think many or any conversations will end with either of these instructions, and I also don't think we'll ever print these instructions out anyways, so I don't think this issue needs any revising right now.

I modified formatting and comments for all files. I certainly could have missed some outdated comments that needed updating, but at the very least it's all in a good place so far.