Tilemap Enable/Disable Experiment
I did a quick experiment to see how the game would be affected when the HUGE tilemap we currently have is enabled or disabled. When it is invisible and I try to enable it, the game actually freezes for a brief second because of how huge the tilemap is.

I think we'll need to subdivide the map into several smaller and separate tilemaps. Each world should have its own tilemap, and I also think that even within those worlds there could be even further subdivided tile maps for better memory usage. We can enable/disable certain ones at certain points based on the Player's location.