Food Buddy FieldState Interface: Prevent User from Selecting 'PLAYER' and 'FUSION' State Within Menu
I added a small amount of code to prevent the user from selecting the 'PLAYER' and 'FUSION' states for their Food Buddies. These states are hard to keep aligned if the Player is able to set them both with number keys and through the interface. I've restricted it so that any FieldState in which the Player is actively using the Food Buddy will be activated just with '1', '2', and '3'. But any independent FieldStates, such as 'FORAGE', 'FOLLOW', and 'SOLO' will be updated through the interface.