Tilemap Ledge Detection and Collision Behavior Experiment: Complete
My first experiment with Tilemap coding was to try and detect when a Character is in-range of a ledge/cliff tile. If so, I wanted to call a specific callback function to process the Character's y-coordinate with respect to that tile. This experiment was useful for helping me get familiar with the TileMapLayer and TileData libraries as well as for getting me comfortable with working with custom data via the editor and in code.

The experiment was successful, but it doesn't seem like adjusting based on y-coordinate is going to be the best way to do this. This form of code causes a jittery effect when the Character's y-coordinate is forcefully reset to stay beneath the ledge tile.

I am going to work on trying typical collision boxes next through the physics layers in the tileset editor.