Tilemap Collision Behavior: Functional But Needs Improvements
I reworked my approach based on the results of the experiment in the previous commit. I figured out how to add physics layers to tiles so they'd collide automatically behind the scenes, but I still need to dig in a bit to figure out how to disable these layers while the Player is jumping.

The Player is currently able to walk around a ledge/cliff terrain with proper collisions and proper z-indexing for perspective. The z-indexing is being done through the use of alternative tiles, which I'm specifically creating for the purpose of perspective. All that the alternate tiles have is a different z-index that will be drawn above the Player, and those alternate tiles should be used whenever the asset is supposed to appear like its in front of or above the Player.

This is working for individual characters, but when multiple are there the code is processing for all of them and updating the tiles differently because of that. The alternative tiles might not be a permanent solution. Maybe I can adjust it so that the Characters have their z-indexes change to fit with the tiles rather than the other way around. That way they adapt to the map, rather than how it currently is with the map adapting to them.